use decimals and store /5 instead of /10

+++ prisma/migrations/20250616183043_tmdb_nypsi_ratings/migration.sql
@@ -7,7 +7,7 @@ CREATE TABLE "tmdbRatings" (
     "type" "RatingType" NOT NULL,
     "id" INTEGER NOT NULL,
     "name" TEXT NOT NULL,
-    "rating" DOUBLE PRECISION NOT NULL,
+    "rating" DECIMAL(2,1) NOT NULL,
 
     CONSTRAINT "tmdbRatings_pkey" PRIMARY KEY ("userId","type","id")
 );
